Move config from server to common
