+++ server/lib/people.js
@@ -1,5 +1,5 @@
+const config = require('@kansa/common/config')
 const { InputError } = require('@kansa/common/errors')
-const config = require('./config')
 const { setKeyChecked } = require('./key')
 const { mailTask, updateMailRecipient } = require('./mail')
 const LogEntry = require('./types/logentry')
