+++ server/lib/user.js
@@ -1,8 +1,8 @@
 const jwt = require('jsonwebtoken')
 const { promisify } = require('util')
 const { matchesId } = require('@kansa/common/auth-user')
+const config = require('@kansa/common/config')
 const { AuthError, InputError } = require('@kansa/common/errors')
-const config = require('./config')
 const { resetExpiredKey } = require('./key')
 const Admin = require('./types/admin')
 const LogEntry = require('./types/logentry')
